requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Queue
redis==5.0.1
nats-py==2.6.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
langchain==0.0.350
langgraph==0.0.20
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.8
bleach==6.1.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
locust==2.17.0

# Code Quality
ruff==0.1.7
black==23.11.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0
click==8.1.7
rich==13.7.0

# Development
pre-commit==3.6.0
ipython==8.17.2
jupyter==1.0.0
